+++ src/models/Person.java
@@ -2,7 +2,7 @@ package models;
 
 
 public abstract class Person {
-    private int id;
+    private Long id;
     private String firstName;
     private String lastName;
     private String email;
@@ -10,18 +10,18 @@ public abstract class Person {
     public Person() {
     }
 
-    public Person(int id, String firstName, String lastName, String email) {
+    public Person(Long id, String firstName, String lastName, String email) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
     }
 
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
